+++ src/app/Fake.DotNet.NuGet/Restore.fs
@@ -1,5 +1,5 @@
 /// Contains tasks which allow to restore NuGet packages from a NuGet package feed like [nuget.org](http://www.nuget.org).
-/// There is also a tutorial about [nuget package restore](../nuget.html) available.
+/// There is also a tutorial about [nuget package restore](/dotnet-nuget.html) available.
 module Fake.DotNet.NuGet.Restore
 
 open Fake.IO