+++ src/app/Fake.DotNet.Paket/Paket.fs
@@ -32,16 +32,12 @@ type PaketPackParams =
       MinimumFromLockFile : bool
       PinProjectReferences : bool }
 
-let private findPaketExecutable () =
-    match Tools.tryFindToolFolderInSubPath "paket" with
-    | Some folder ->
-        folder @@ "paket"
-    | None ->
-        (Tools.findToolFolderInSubPath "paket.exe" (Directory.GetCurrentDirectory() @@ ".paket")) @@ "paket.exe"
+let internal findPaketExecutable (baseDir) =
+    ProcessUtils.findLocalTool "PAKET" "paket" [ Path.Combine(baseDir, ".paket") ]
 
 /// Paket pack default parameters
 let PaketPackDefaults() : PaketPackParams =
-    { ToolPath = findPaketExecutable ()
+    { ToolPath = findPaketExecutable ""
       TimeOut = TimeSpan.FromMinutes 5.
       Version = null
       SpecificVersions = []
@@ -71,7 +67,7 @@ type PaketPushParams =
 
 /// Paket push default parameters
 let PaketPushDefaults() : PaketPushParams =
-    { ToolPath = findPaketExecutable ()
+    { ToolPath = findPaketExecutable ""
       TimeOut = System.TimeSpan.MaxValue
       PublishUrl = null
       EndPoint =  null
@@ -91,7 +87,7 @@ type PaketRestoreParams =
 
 /// Paket restore default parameters
 let PaketRestoreDefaults() : PaketRestoreParams =
-    { ToolPath = findPaketExecutable ()
+    { ToolPath = findPaketExecutable ""
       TimeOut = System.TimeSpan.MaxValue
       WorkingDir = "."
       ForceDownloadOfPackages = false